// src/data/enums.ts
export enum FontSize {
  XS = "text-xs",
  SM = "text-sm",
  BASE = "text-base",
  LG = "text-lg",
  XL = "text-xl",
  XL2 = "text-2xl",
  XL3 = "text-3xl",
  XL4 = "text-4xl",
  XL5 = "text-5xl",
  XL6 = "text-6xl",
  XL7 = "text-7xl",
  XL8 = "text-8xl",
  XL9 = "text-9xl",
}

export enum FontWeight {
    THIN = "font-thin",
    EXTRALIGHT = "font-extralight",
    LIGHT = "font-light",
    NORMAL = "font-normal",
    MEDIUM = "font-medium",
    SEMIBOLD = "font-semibold",
    BOLD = "font-bold",
    EXTRABOLD = "font-extrabold",
    BLACK = "font-black",
}